+++ Maya3D/Classes/OpenGL/GLTexture.m
@@ -77,7 +77,7 @@
 	// Specify a 2D texture image, providing the a pointer to the image data in memory
 	// void glTexImage2D (GLenum target, GLint level, GLint internalformat, GLsizei width, GLsizei height, GLint border, GLenum format, GLenum type, const GLvoid *pixels);
 	if (1)
-		glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, width, height, 0, GL_RGBA, GL_UNSIGNED_BYTE, texData);
+		glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, (int)width, (int)height, 0, GL_RGBA, GL_UNSIGNED_BYTE, texData);
 	else
 	{
 		// Remove alpha data
